[Step1] refactor : 패키지 수정
