--- src/main/kotlin/resume/Language.kt
@@ -1,3 +1,0 @@
-package resume
-
-data class Language(val name: String, val level: Int)
