--- src/main/kotlin/resume/Skill.kt
@@ -1,8 +1,0 @@
-package resume
-
-sealed interface Skill {
-    val name: String
-
-    data class Hard(override val name: String) : Skill
-    data class Soft(override val name: String) : Skill
-}
